tasl

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -5,7 +5,7 @@
 */
 void print_python_list_info(PyObject *p)
 {
-    int size_of_list, alloc, i;
+    long int size_of_list, alloc, i;
     PyObject *j;
 
     size_of_list = Py_SIZE(p);
